requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.1
gunicorn==20.0.4
djangorestframework==3.12.4
python-dotenv==0.18.0
django-cors-headers==3.7.0
psycopg2-binary==2.9.3